debate-bot/requirements.txt
python-telegram-bot==21.6
google-genai==1.7.0
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
# html parser backend for BeautifulSoup (pure Python, works on ARM)
lxml==5.3.0
